Add files via upload
This is a machine learning workflow that uses Support Vector Machines (SVM) with cross-validation to evaluate classification performance. The dataset includes  E.coli samples from different sources and its resistance to 6 different antibiotics. Samples showing resistance are denoted as 1 and the ones which are not resistant are denoted as 0. Samples showing MDR are denoted as 1